fix: remove deprecated `int()` method from schema

+++ src/lib/orpc/schemas/shared.ts
@@ -1,8 +1,8 @@
 import { z } from "zod/v4";
 
 export const paginationSchema = z.object({
-	pageSize: z.coerce.number().int().min(1).max(100).default(20),
-	pageIndex: z.coerce.number().int().min(0).default(0),
+	pageSize: z.coerce.number().min(1).max(100).default(20),
+	pageIndex: z.coerce.number().min(0).default(0),
 });
 
 export const statusSchema = z.object({
